Add files via upload
Main game mechanics are done, if click on a cell with no neighbors it will reveal surrounding cells that don't have neighbors. Color added to each number depending on what number it is (ex. "1" is blue, "2" is green, etc.). Lots of debugging to do. Current bugs and missing features include:
- revealNeighbors reveals cells that are flagged
- unflagging doesn't work
- player should be able to click on a cell and reveal its neighbors if the cell clicked is already revealed 
- End game function needs to be added
- GUI, beautify everything, cleanup code/optimize (all this will be done when most/all bugs are fixed)